# DESIGNdOCS/openapi.yaml
openapi: 3.0.3
info:
  title: Welo Data Annotation Platform API
  description: |
    RESTful API for the Welo Data Annotation Platform — a microservices monorepo.
    Each service runs independently on its own port. All services share the prefix `/api/v1`
    and accept/return JSON.

    ## Authentication
    All protected endpoints require a JWT Bearer token:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Microservice Ports
    | Service | Port |
    |---------|------|
    | Workflow Engine | 3001 |
    | Auth Service | 3002 |
    | Task Management | 3003 |
    | Project Management | 3004 |
    | Annotation QA Service | 3005 |
    | File Storage Service | 3006 |
    | Export Service | 3007 |
    | Notification Service | 3008 |
    | Analytics Service | 3009 |
    | Audit Service | 3010 |

    In production all services sit behind a reverse proxy / API gateway.
    Swagger UI is available at `http://localhost:{port}/api` for each service.
  version: 1.0.0
  contact:
    name: Welo Platform Team
servers:
  - url: http://localhost:3001/api/v1
    description: Workflow Engine (local)
  - url: http://localhost:3002/api/v1
    description: Auth Service (local)
  - url: http://localhost:3003/api/v1
    description: Task Management (local)
  - url: http://localhost:3004/api/v1
    description: Project Management (local)
  - url: http://localhost:3005/api/v1
    description: Annotation QA Service (local)
  - url: http://localhost:3006/api/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /files:
    get:
      tags: [files]
      summary: List files for a project/batch
      operationId: listFiles
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: batchId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFiles'
  /files/upload:
    post:
      tags: [files]
      summary: Direct file upload (development / local storage mode)
      operationId: uploadFile
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: batchId
          in: query
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, projectId]
              properties:
                file:
                  type: string
                  format: binary
                projectId:
                  type: string
                  format: uuid
                batchId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileRecord'
  /files/presigned-url:
    post:
      tags: [files]
      summary: Request a presigned S3 PUT URL for client-side upload (production)
      operationId: requestPresignedUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignedUrlDto'
      responses:
        '201':
          description: Presigned URL returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                    format: uri
                  fileId:
                    type: string
                    format: uuid
                  fileKey:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /files/{id}:
    get:
      tags: [files]
      summary: Get file metadata by ID
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: File metadata retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileRecord'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [files]
      summary: Delete a file (soft delete, publishes file.deleted event)
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: File deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{id}/download:
    get:
      tags: [files]
      summary: Get presigned download URL for a file
      operationId: getFileDownloadUrl
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Download URL returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{id}/confirm:
    post:
      tags: [files]
      summary: Confirm upload complete after client-side S3 PUT
      operationId: confirmUpload
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileSize]
              properties:
                fileSize:
                  type: integer
                  description: File size in bytes
      responses:
        '200':
          description: File record confirmed as READY
        '404':
          $ref: '#/components/responses/NotFound'

  # =========================================================================
  # EXPORT SERVICE  (port 3007)
  # =========================================================================
  /exports:
    get:
      tags: [exports]
      summary: List exports, optionally filtered by project/batch/status
      operationId: listExports
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: batchId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ExportStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Exports retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExports'
    post:
      tags: [exports]
      summary: Request a new export job
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportDto'
      responses:
        '201':
          description: Export job queued (status=PENDING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportRecord'
  /exports/{id}:
    get:
      tags: [exports]
      summary: Get export status by ID
      operationId: getExport
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Export retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportRecord'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [exports]
      summary: Cancel an export (idempotent for already-completed exports)
      operationId: cancelExport
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Export cancelled
        '404':
          $ref: '#/components/responses/NotFound'
  /exports/{id}/download:
    get:
      tags: [exports]
      summary: Get download URL for a completed export
      operationId: getExportDownloadUrl
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Download URL returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
        '404':
          $ref: '#/components/responses/NotFound'
  /exports/{id}/retry:
    post:
      tags: [exports]
      summary: Retry a failed export job
      operationId: retryExport
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Export requeued for processing
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # =========================================================================
  # NOTIFICATION SERVICE  (port 3008)
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
  responses:
    BadRequest:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: The requested resource was not found
            details:
              type: object
              additionalProperties: true
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5

    # -------------------------------------------------------------------------
    # Auth / Users
    # -------------------------------------------------------------------------
...
    TaskStatus:
      type: string
      enum:
        - QUEUED
        - ASSIGNED
        - IN_PROGRESS
        - SUBMITTED
        - IN_REVIEW
        - APPROVED
        - REJECTED
        - REWORK
        - COMPLETED
        - CANCELLED
      example: QUEUED
...
    FileStatus:
      type: string
      enum: [PENDING, UPLOADING, READY, FAILED, DELETED]
    FileRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
          description: Size in bytes
        fileKey:
          type: string
          description: Storage key / path
        status:
          $ref: '#/components/schemas/FileStatus'
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    PaginatedFiles:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            files:
              type: array
              items:
                $ref: '#/components/schemas/FileRecord'
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
    PresignedUrlDto:
      type: object
      required: [fileName, mimeType, projectId]
      properties:
        fileName:
          type: string
          example: image_001.jpg
        mimeType:
          type: string
          example: image/jpeg
        projectId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid

    # -------------------------------------------------------------------------
    # Exports
    # -------------------------------------------------------------------------
    ExportStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
    ExportFormat:
      type: string
      enum: [JSON, CSV, COCO, PASCAL_VOC, YOLO, CUSTOM]
    ExportRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        format:
          $ref: '#/components/schemas/ExportFormat'
        status:
          $ref: '#/components/schemas/ExportStatus'
        fileUrl:
          type: string
          format: uri
          nullable: true
        requestedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    PaginatedExports:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            exports:
              type: array
              items:
                $ref: '#/components/schemas/ExportRecord'
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
    CreateExportDto:
      type: object
      required: [projectId, format]
      properties:
        projectId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        format:
          $ref: '#/components/schemas/ExportFormat'
        filters:
          type: object
          properties:
            status:
              $ref: '#/components/schemas/TaskStatus'
            dateFrom:
              type: string
              format: date
            dateTo:
              type: string
              format: date
        includeRejected:
          type: boolean
          default: false

    # -------------------------------------------------------------------------
    # Notifications
    # -------------------------------------------------------------------------
...